+++ lib/file-io/index.js
