+++ lib/file-io/test/file-io.js
@@ -1,14 +1,14 @@
 const should = require("chai").should();
 
-const file = require("../src/file-io");
+const io = require("../");
 
 describe("file-io", () => {
     describe("read", () => {
         it("should open files matching glob", (done) => {
-            file.open("**/*.js").then(files => {
+            io.open("**/*.js").then(files => {
                 files.should.exist;
                 done();
-            })
+            });
         });
     });
 });
