+++ lib/mapper/test/mapper.js
@@ -1,6 +1,6 @@
 var should = require('chai').should();
 
-var mapper = require('../src/mapper');
+var mapper = require('../');
 
 describe('map', () => {
     
@@ -17,6 +17,7 @@ describe('map', () => {
             }.should.throw(expectedError + typeof "");
             
             mapper.map([]).should.be.ok;
+            
         });
     });
