+++ lib/ruin.js
