+++ lib/short-css-class-names/char-range/index.js
