+++ lib/short-css-class-names/char-range/test/char-range.js
@@ -1,6 +1,6 @@
 var should = require('chai').should();
 
-var charRange = require('../src/char-range');
+var charRange = require('../');
 
 describe('char-range', function() {
   describe('create', function () {
