+++ lib/short-css-class-names/index.js
@@ -12,7 +12,7 @@
 module.exports = function(rangeLength) {
     "use strict";
     
-    let charRange = require("../char-range"),
+    let charRange = require("./char-range"),
         digits = charRange("0", "9"),
         letters = charRange("A", "Z").concat(charRange("a", "z")),
         hyphenAndUnderscore = ["-", "_"];
