+++ lib/short-css-class-names/test/short-css-class-names.js
@@ -7,7 +7,7 @@
         const maxLength = 216372;
         
         describe("defaults", function() {
-            const classNames = require("../src/short-css-class-names")();
+            const classNames = require("../")();
             
             it("should contain heaps of names", function() {
                 classNames.length.should.equal(maxLength);
@@ -24,13 +24,13 @@
         
         describe("specified length", function() {
             it("exceeded range should return default max", function() {
-                const classNames = require("../src/short-css-class-names")(Number.MAX_VALUE);                
+                const classNames = require("../")(Number.MAX_VALUE);                
                 classNames.length.should.equal(maxLength);
             });
             
             it("exceeded range should return default max", function() {
                 const rangeLength = 999;
-                const classNames = require("../src/short-css-class-names")(rangeLength);                
+                const classNames = require("../")(rangeLength);                
                 classNames.length.should.equal(rangeLength);
             });
         });
